fix: refactor home and seleted business unit pages
